extract link feed detection
